+++ cmd/commandfuncs.go
@@ -269,10 +269,6 @@ func cmdRun(fl Flags) (int, error) {
 		}
 	}
 
-	if err := NotifyReadiness(); err != nil {
-		caddy.Log().Error("unable to notify readiness to service manager", zap.Error(err))
-	}
-
 	select {}
 }
 
@@ -293,15 +289,6 @@ func cmdReload(fl Flags) (int, error) {
 	reloadCmdConfigAdapterFlag := fl.String("adapter")
 	reloadCmdAddrFlag := fl.String("address")
 	reloadCmdForceFlag := fl.Bool("force")
-
-	if err := NotifyReloading(); err != nil {
-		caddy.Log().Error("unable to notify reloading to service manager", zap.Error(err))
-	}
-	defer func() {
-		if err := NotifyReadiness(); err != nil {
-			caddy.Log().Error("unable to notify readiness to service manager", zap.Error(err))
-		}
-	}()
 
 	// get the config in caddy's native format
 	config, configFile, err := loadConfig(reloadCmdConfigFlag, reloadCmdConfigAdapterFlag)
